requirements.txt
datetime
newsapi-python
TextBlob
matplotlib
numpy
pandas
oandapyV20
psycopg2
vaderSentiment
python-twitter